+++ bricks/replay_cubit/__brick__/{{name.snakeCase()}}_cubit.dart
@@ -1,0 +1,8 @@
+import 'package:replay_bloc/replay_bloc.dart';
+
+part '{{name.snakeCase()}}_event.dart';
+part '{{name.snakeCase()}}_state.dart';
+
+class {{name.pascalCase()}}Cubit extends ReplayCubit<{{name.pascalCase()}}State> {
+  {{name.pascalCase()}}Cubit() : super(const {{name.pascalCase()}}State());
+}