+++ examples/flutter_firebase_login/packages/authentication_repository/test/authentication_repository_test.dart
@@ -104,7 +104,7 @@ void main() {
     });
 
     test('creates FirebaseAuth instance internally when not injected', () {
-      expect(() => AuthenticationRepository(), isNot(throwsException));
+      expect(AuthenticationRepository.new, isNot(throwsException));
     });
 
     group('signUp', () {
@@ -275,7 +275,7 @@ void main() {
     group('logOut', () {
       test('calls signOut', () async {
         when(() => firebaseAuth.signOut()).thenAnswer((_) async {});
-        when(() => googleSignIn.signOut()).thenAnswer((_) async {});
+        when(() => googleSignIn.signOut()).thenAnswer((_) async => null);
         await authenticationRepository.logOut();
         verify(() => firebaseAuth.signOut()).called(1);
         verify(() => googleSignIn.signOut()).called(1);