+++ extensions/vscode/src/utils/analyze-dependencies.ts
@@ -1,8 +1,10 @@
 import * as _ from "lodash";
 import * as semver from "semver";
 import { window, env, Uri } from "vscode";
-import { getLatestPackageVersion, getPubspec } from ".";
+import { getLatestPackageVersion, getPubspecLock } from ".";
 import { updatePubspecDependency } from "./update-pubspec-dependency";
+
+const DEFAULT_VERSION_VALUE = "0.0.0";
 
 interface Dependency {
   name: string;
@@ -48,12 +50,12 @@ const devDeps = [{ name: "bloc_test", actions: [openBlocMigrationGuide] }];
 export async function analyzeDependencies() {
   const dependencies = await getDependencies(deps);
   const devDependencies = await getDependencies(devDeps);
-  const pubspec = await getPubspec();
-  const pubspecDependencies = _.get(pubspec, "dependencies", {});
-  const pubspecDevDependencies = _.get(pubspec, "dev_dependencies", {});
+  const pubspecLock = await getPubspecLock();
 
-  checkForUpgrades(dependencies, pubspecDependencies);
-  checkForUpgrades(devDependencies, pubspecDevDependencies);
+  const pubspecLockDependencies = _.get(pubspecLock, "packages", {});
+
+  checkForUpgrades(dependencies, pubspecLockDependencies);
+  checkForUpgrades(devDependencies, pubspecLockDependencies);
 }
 
 function checkForUpgrades(
@@ -64,43 +66,47 @@ function checkForUpgrades(
     const dependency = dependencies[i];
     if (_.isEmpty(dependency.version)) continue;
     if (_.has(pubspecDependencies, dependency.name)) {
-      const dependencyVersion = _.get(
+      const currentDependencyVersion = _.get(
         pubspecDependencies,
         dependency.name,
-        "latest"
-      );
-      if (dependencyVersion === "latest") continue;
-      if (dependencyVersion === "any") continue;
-      if (dependencyVersion == null) continue;
-      if (typeof dependencyVersion !== "string") continue;
-      const minVersion = _.get(
-        semver.minVersion(dependencyVersion),
-        "version",
-        "0.0.0"
-      );
-      if (!semver.satisfies(minVersion, dependency.version)) {
-        window
-          .showWarningMessage(
-            `This workspace contains an outdated version of ${dependency.name}. Please update to ${dependency.version}.`,
-            ...dependency.actions.map((action) => action.name).concat("Update")
-          )
-          .then((invokedAction) => {
-            if (invokedAction === "Update") {
-              return updatePubspecDependency({
-                name: dependency.name,
-                latestVersion: `^${dependency.version}`,
-                currentVersion: dependencyVersion,
-              });
-            }
-            const action = dependency.actions.find(
-              (action) => action.name === invokedAction
-            );
-            if (!_.isNil(action)) {
-              action.callback();
-            }
-          });
-      }
+      ).version;
+      
+      const hasLatestVersion = currentDependencyVersion === dependency.version;
+      if (hasLatestVersion) continue;
+      
+      showUpdateMessage(dependency, currentDependencyVersion);
     }
+  }
+}
+
+function showUpdateMessage(dependency : Dependency, dependencyVersion : string) {
+  const minVersion = _.get(
+    semver.minVersion(dependencyVersion),
+    "version",
+    DEFAULT_VERSION_VALUE
+  );
+  
+  if (!semver.satisfies(minVersion, dependency.version)) {
+    window
+      .showWarningMessage(
+        `This workspace contains an outdated version of ${dependency.name}. Please update to ${dependency.version}.`,
+        ...dependency.actions.map((action) => action.name).concat("Update")
+      )
+      .then((invokedAction) => {
+        if (invokedAction === "Update") {
+          return updatePubspecDependency({
+            name: dependency.name,
+            latestVersion: `^${dependency.version}`,
+            currentVersion: dependencyVersion,
+          });
+        }
+        const action = dependency.actions.find(
+          (action) => action.name === invokedAction
+        );
+        if (!_.isNil(action)) {
+          action.callback();
+        }
+      });
   }
 }